feat(workflows): add loops (#1001)
<!-- Please make sure there is an issue that this PR is correlated to. -->

## Changes

<!-- If there are frontend changes, please include screenshots. -->

+++ lib/chirp-workflow/core/src/activity.rs
@@ -19,7 +19,7 @@ pub trait Activity {
 	type Output: Serialize + DeserializeOwned + Debug + Send;
 
 	const NAME: &'static str;
-	const MAX_RETRIES: u32;
+	const MAX_RETRIES: usize;
 	const TIMEOUT: std::time::Duration;
 
 	async fn run(ctx: &ActivityCtx, input: &Self::Input) -> GlobalResult<Self::Output>;
